00:36

+++ src/tasks/tasks.service.ts
@@ -1,3 +1,4 @@
+import { User } from './../auth/user.entity';
 import { TaskRepository } from './task.repository';
 import { GetTasksFilterDto } from './dto/getTasks-filter.dto';
 import { CreateTaskDto } from './dto/createTask.dto';
@@ -13,36 +14,40 @@ export class TasksService {
   constructor(
     @InjectRepository(TaskRepository)
     private taskRepository: TaskRepository,
-  ){}
-  
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return await this.taskRepository.getTasks(filterDto)
+  ) {}
+
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return await this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-      const found = await this.taskRepository.findOne(id)
-      
-      if(!found) {
-        throw new NotFoundException(`Task with ID "${id}" not found`)
-      }
-      return found
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({where: {id, userId: user.id}});
+
+    if (!found) {
+      throw new NotFoundException(`Task with ID "${id}" not found`);
     }
-
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+    return found;
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id)
-    if(result.affected === 0 ){
-      throw new NotFoundException(`Task with ID "${id}" not found`)
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
+  }
+
+  async deleteTask(
+    id: number,
+    user: User,
+  ): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: +user.id });
+
+    if (result.affected === 0) {
+      throw new NotFoundException(`Task with ID "${id}" not found`);
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id)
-    task.status = status
-    await task.save()
-    return task
+  async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
+    task.status = status;
+    await task.save();
+    return task;
   }
 }
